update line chart options to import from chart options file
