+++ src/components/Databases.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import chartOptions from '../helpers/ChartOptions';
 
@@ -15,46 +14,12 @@ class Databases extends React.Component {
       ]
     };
 
-    const options = {
-      fullWidth: true,
-      height: '300px',
-      chartPadding: {
-        right: 40,
-        left: 20
-      },
-      plugins: [
-        Chartist.plugins.legend({
-          legendNames: ['PostgreSQL', 'MySQL', 'Other'],
-        }),
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-        Chartist.plugins.ctAxisTitle({
-          axisY: {
-            axisTitle: 'Percent',
-            axisClass: 'ct-axis-title',
-            textAnchor: 'middle',
-            offset: {
-              x: 0,
-              y: -1
-            },
-            flipTitle: false,
-          }
-        })
-      ]
-    };
-
-
-
     return (
       <div>
         <ChartistGraph
           data={data}
           type={'Line'}
-          options={options}
+          options={chartOptions.lineOptions}
           responsiveOptions={chartOptions.lineResponsiveOptions}/>
       </div>
     );
