+++ src/components/RailsServers.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import chartOptions from '../helpers/ChartOptions';
 
@@ -15,52 +14,13 @@ class RailsServers extends React.Component {
       ]
     };
 
-    const options = {
-      fullWidth: true,
-      height: '300px',
-      low: 0,
-      chartPadding: {
-        right: 40,
-        left: 20
-      },
-      plugins: [
-        Chartist.plugins.legend({
-          legendNames: [
-            'Puma',
-            'Unicorn',
-            'Passenger'
-          ],
-          clickable: false,
-        }),
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-        Chartist.plugins.ctAxisTitle({
-          axisY: {
-            axisTitle: 'Percent',
-            axisClass: 'ct-axis-title',
-            textAnchor: 'middle',
-            offset: {
-              x: 0,
-              y: -1
-            },
-            flipTitle: false,
-          }
-        })
-      ]
-    };
-
-
     return (
       <div>
         <ChartistGraph
           data={data}
           type={'Line'}
           responsiveOptions={chartOptions.lineResponsiveOptions}
-          options={options} />
+          options={chartOptions.lineOptions} />
       </div>
     );
   }
